// nestjs/src/rents/spreadsheets/models/spreadsheets.model.ts
export interface Sheet {
    sheetId: number;
    title: string;
    rows: { value: string | number, backgroundColor?: string }[][];
}

export interface SpreadSheet {
    id: string;
    sheets: Sheet[];
    title: string;
}

export interface SpreadSheetUpdate {
    sheetTitle: string,
    cell: string;
    backgroundColor: string;
    value: string | number;
}